fix #291261 Volta Mouse Drop default first staff
Previously, dragging a volta from the palette and dropping onto a measure would cause the volta to only be applied to the particular staff under the mouse. However this causes problems because MuseScore generally assumes that first staff is the authority on voltas when creating and dealing with part scores.

This PR makes volta drops only be applied to the first staff, unless the user holds control while dropping, in which case the previous behavior of the volta gets applied to the particular staff will occur.